+++ graphite_beacon/handlers/__init__.py
@@ -64,3 +64,4 @@ from .slack import SlackHandler          # noqa
 from .smtp import SMTPHandler            # noqa
 from .cli import CliHandler              # noqa
 from .opsgenie import OpsgenieHandler    # noqa
+from .victorops import VictorOpsHandler  # noqa